Deprecate last old test module and rename newer ones
Fixes #29
